temp_requirements.txt
asgiref==3.8.1
channels==4.2.2
channels-redis==4.2.1
django-crispy-forms==2.4
django-widget-tweaks==1.5.0
pillow==11.2.1
python-dotenv==1.1.0
